It didn't delete the journal entry, fixed

+++ app/firebase/deleteJournalEntryFromFirebase.jsx
@@ -1,24 +1,21 @@
 import { getFirestore, doc, getDoc, deleteDoc } from "firebase/firestore";
-import { getStorage, ref, deleteObject } from "firebase/storage";
 import { Alert } from "react-native";
 
 export const deleteJournalEntryFromFirebase = async (entryId) => {
   try {
     const db = getFirestore();
-    const storage = getStorage();
 
     const entryRef = doc(db, "JournalEntries", entryId);
     const entrySnapshot = await getDoc(entryRef);
-    const image = entrySnapshot.data().Images;
 
-    // delete image from storage
-    const imageRef = ref(storage, image);
-    await deleteObject(imageRef);
+    // making sure the actual entry exists
+    if (entrySnapshot.exists()) {
+      await deleteDoc(entryRef);
 
-    // delete document
-    await deleteDoc(entryRef);
-    
-    Alert.alert('Entry Deleted');
+      Alert.alert('Entry Deleted');
+    } else {
+      Alert.alert('Entry does not exist');
+    }
   } catch (error) {
     console.error('Error deleting journal entry:', error);
     throw error;
